+++ veolia_api/__init__.py
@@ -14,4 +14,4 @@ __all__ = [
     "resolve_portal_url",
 ]
 
-__version__ = "2.4.3"
+__version__ = "2.4.4"